move out more shared code
