+++ packages/core/src/test/amazonqFeatureDev/utils.ts
@@ -20,6 +20,8 @@ import { VirtualMemoryFile } from '../../shared/virtualMemoryFile'
 import path from 'path'
 import { featureDevScheme } from '../../amazonqFeatureDev/constants'
 import { HttpResponse } from 'aws-sdk'
+import { DefaultCodeWhispererClient } from '../../codewhisperer'
+import { stub } from '../utilities/stubber'
 
 export function createMessenger(): Messenger {
     return new Messenger(
@@ -119,42 +121,6 @@ export async function createController(): Promise<ControllerSetup> {
     }
 }
 
-export const mockCreateCodeScanResponse = {
-    $response: {
-        data: {
-            jobId: 'jobId',
-            status: 'Pending',
-        },
-        requestId: 'requestId',
-        hasNextPage: () => false,
-        error: undefined,
-        nextPage: () => undefined,
-        redirectCount: 0,
-        retryCount: 0,
-        httpResponse: new HttpResponse(),
-    },
-    jobId: 'jobId',
-    status: 'Pending',
-}
-
-export const mockCreateUploadUrlResponse = {
-    $response: {
-        data: {
-            uploadId: 'uploadId',
-            uploadUrl: 'uploadUrl',
-        },
-        requestId: 'requestId',
-        hasNextPage: () => false,
-        error: undefined,
-        nextPage: () => undefined,
-        redirectCount: 0,
-        retryCount: 0,
-        httpResponse: new HttpResponse(),
-    },
-    uploadId: 'uploadId',
-    uploadUrl: 'https://test.com',
-}
-
 export const mockGetCodeScanResponse = {
     $response: {
         data: {
@@ -171,44 +137,89 @@ export const mockGetCodeScanResponse = {
     status: 'Completed',
 }
 
-export const mockCodeScanFindings = JSON.stringify([
-    {
-        filePath: 'workspaceFolder/python3.7-plain-sam-app/hello_world/app.py',
-        startLine: 1,
-        endLine: 1,
-        title: 'title',
-        description: {
-            text: 'text',
-            markdown: 'markdown',
-        },
-        detectorId: 'detectorId',
-        detectorName: 'detectorName',
-        findingId: 'findingId',
-        relatedVulnerabilities: [],
-        severity: 'High',
-        remediation: {
-            recommendation: {
-                text: 'text',
-                url: 'url',
-            },
-            suggestedFixes: [],
-        },
-        codeSnippet: [],
-    } satisfies model.RawCodeScanIssue,
-])
+export function createMockClient() {
+    const mockClient = stub(DefaultCodeWhispererClient)
 
-export const mockListCodeScanFindingsResponse = {
-    $response: {
-        data: {
-            codeScanFindings: mockCodeScanFindings,
+    const mockCreateCodeScanResponse = {
+        $response: {
+            data: {
+                jobId: 'jobId',
+                status: 'Pending',
+            },
+            requestId: 'requestId',
+            hasNextPage: () => false,
+            error: undefined,
+            nextPage: () => undefined,
+            redirectCount: 0,
+            retryCount: 0,
+            httpResponse: new HttpResponse(),
         },
-        requestId: 'requestId',
-        hasNextPage: () => false,
-        error: undefined,
-        nextPage: () => undefined,
-        redirectCount: 0,
-        retryCount: 0,
-        httpResponse: new HttpResponse(),
-    },
-    codeScanFindings: mockCodeScanFindings,
+        jobId: 'jobId',
+        status: 'Pending',
+    }
+    const mockCreateUploadUrlResponse = {
+        $response: {
+            data: {
+                uploadId: 'uploadId',
+                uploadUrl: 'uploadUrl',
+            },
+            requestId: 'requestId',
+            hasNextPage: () => false,
+            error: undefined,
+            nextPage: () => undefined,
+            redirectCount: 0,
+            retryCount: 0,
+            httpResponse: new HttpResponse(),
+        },
+        uploadId: 'uploadId',
+        uploadUrl: 'https://test.com',
+    }
+
+    const mockCodeScanFindings = JSON.stringify([
+        {
+            filePath: 'workspaceFolder/python3.7-plain-sam-app/hello_world/app.py',
+            startLine: 1,
+            endLine: 1,
+            title: 'title',
+            description: {
+                text: 'text',
+                markdown: 'markdown',
+            },
+            detectorId: 'detectorId',
+            detectorName: 'detectorName',
+            findingId: 'findingId',
+            relatedVulnerabilities: [],
+            severity: 'High',
+            remediation: {
+                recommendation: {
+                    text: 'text',
+                    url: 'url',
+                },
+                suggestedFixes: [],
+            },
+            codeSnippet: [],
+        } satisfies model.RawCodeScanIssue,
+    ])
+
+    const mockListCodeScanFindingsResponse = {
+        $response: {
+            data: {
+                codeScanFindings: mockCodeScanFindings,
+            },
+            requestId: 'requestId',
+            hasNextPage: () => false,
+            error: undefined,
+            nextPage: () => undefined,
+            redirectCount: 0,
+            retryCount: 0,
+            httpResponse: new HttpResponse(),
+        },
+        codeScanFindings: mockCodeScanFindings,
+    }
+
+    mockClient.createCodeScan.resolves(mockCreateCodeScanResponse)
+    mockClient.createUploadUrl.resolves(mockCreateUploadUrlResponse)
+    mockClient.getCodeScan.resolves(mockGetCodeScanResponse)
+    mockClient.listCodeScanFindings.resolves(mockListCodeScanFindingsResponse)
+    return mockClient
 }
